A class that can deserialize a Java bin files.
What this function returns and its usefulness strongly depends
on the knowledge of exactly how is the binary file built.
The reader fetches blocks of bytes from the file in memory, and
subsequent calls to read_int() etc reads from the block currently
in memory.
When the end of the block is reached while trying to read some values,
an exception is thrown, caught, and used to fetch another block from
the file.